rag_server/requirements.txt
fastapi
uvicorn
pymongo
chromadb
sentence-transformers
google-generativeai
python-dotenv
requests
